:fire: refactor : change BigDecimal to int or Integer

+++ com.templlo.service.promotion/src/main/java/com/templlo/service/common/dto/DetailProgramResponse.java
@@ -1,13 +1,12 @@
 package com.templlo.service.common.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 public record DetailProgramResponse(
 	UUID programId,
 	String type,
-	BigDecimal programFee,
+	int programFee,
 	LocalDate programDate
 ) {
 }
